Better logging, minor bugs fixed.
The tests now give some progress reports on console.

Non-checked executions (typically for cleaning up server state)
are categorised as either expect success or maybe fail.  If an
expect success execution fails then the output is logged to console.
A maybe-fail execution only shows the RC.

The output generated by the test is now cleared; only the errors for
a given executable are shown for that executable.

Some errors in G2 tests are fixed: wrong URI and attempting to
delete a directory with srmrm.