rpm: enforce SL5 compatibility when building RPM packages
A recent security update broke compatibility with SL5 by changing the
default compression from gzip (DEFLATE) to xz (LZMA/-2) and changing
the hash from MD5 to SHA256.

These additional requirements are evident from additional requirements
from the package.

The `rpm -qRp` command shows dependencies; when build in SL6 machines,
the following two additional dependencies are included:

    rpmlib(FileDigests) <= 4.6.0-1
    rpmlib(PayloadIsXz) <= 5.2-1

The version of rpm provided with SL5 does not provide the
`rpmlib(FileDigests)` or the `rpmlib(PayloadIsXz)` capabilities,
preventing the RPM from being installed.

This patch adds definitions so that the former defaults (gzip, MD5) are
enforced.  The result is RPM files that may be installed on either SL5
or SL6.