Add dCache version probe via WebDAV for newer dCache versions.
The info probe is proving slow and problematic as it sometimes fails.
If it fails, a dCache v2.10 is assumed, which enables SRM v1 tests.

The SRMv1 tests will fail for 2.15 and later, as these versions do
not support SRMv1, however, these versions also provide dCache
version as an HTTP header in responses from WebDAV door.

Therefore, we first try the WebDAV door; if this fails (because
the dCache version is earlier than 2.13) then the code tries
to use info.  If this fails, then the default dCache version is
safe.

Overall, we have faster, and more likely to succeed in detecting
dCache version, with the most likely failures happening in a safe
fashion.